Additonal safeguard : Prefer the use of unsigned pointer sizes for reprensenting length in core functions instead of signed int in order to prevent possibly existing overflows for crafted requests. This should also fix buffer overflows in the case someone run that software on a 16 bit x86 with an already supported ᴏꜱ.

+++ src/src/auths/check_serv_cond.c
@@ -72,7 +72,7 @@ HDEBUG(D_auth)
       debug_printf("  $auth%d = %s\n", i + 1, auth_vars[i]);
     }
   for (i = 1; i <= expand_nmax; i++)
-    debug_printf("  $%d = %.*s\n", i, expand_nlength[i], expand_nstring[i]);
+    debug_printf("  $%u = %.*s\n", i, (unsigned int)expand_nlength[i], expand_nstring[i]);
   debug_print_string(ablock->server_debug_string);    /* customized debug */
   }
